refactored the database
