+++ lib/models/language.dart
@@ -1,7 +1,7 @@
-import 'package:pokemon_pokedex/models/database/constants.dart';
 import 'package:pokemon_pokedex/models/utility/deserializable.dart';
 import 'package:pokemon_pokedex/models/utility/model.dart';
 import 'package:pokemon_pokedex/models/utility/serializable.dart';
+import 'package:pokemon_pokedex/resources/database/sqflite/constants.dart';
 import 'package:to_string/to_string.dart';
 
 part 'language.g.dart';
