+++ lib/models/utility/pokemon_base_information.dart
@@ -1,9 +1,9 @@
-import 'package:pokemon_pokedex/models/database/constants.dart';
 import 'package:pokemon_pokedex/models/language.dart';
 import 'package:pokemon_pokedex/models/pokemon/pokemon_type.dart';
 import 'package:pokemon_pokedex/models/utility/deserializable.dart';
 import 'package:pokemon_pokedex/models/utility/model.dart';
 import 'package:pokemon_pokedex/models/utility/serializable.dart';
+import 'package:pokemon_pokedex/resources/database/sqflite/constants.dart';
 import 'package:to_string/to_string.dart';
 
 part 'pokemon_base_information.g.dart';
