+++ lib/screens/drawer/main_drawer.dart
@@ -1,14 +1,17 @@
 import 'package:flutter/material.dart';
-import 'package:pokemon_pokedex/models/database/base_information_database.dart';
 import 'package:pokemon_pokedex/models/language.dart';
+import 'package:pokemon_pokedex/models/utility/pokemon_base_information.dart';
+import 'package:pokemon_pokedex/resources/database/storage_provider.dart';
 import 'package:pokemon_pokedex/screens/drawer/widgets/drawer_tile.dart';
-import 'package:pokemon_pokedex/screens/searchScreen/search_screen.dart';
 import 'package:pokemon_pokedex/utils/constants.dart';
 import 'package:pokemon_pokedex/utils/routes.dart';
+import 'package:provider/provider.dart';
 
 class MainDrawer extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    final StorageProvider database = Provider.of<StorageProvider>(context);
+
     final List<DrawerTile> drawerTileList = <DrawerTile>[
       DrawerTile(
         title: "Pokédex",
@@ -31,43 +34,29 @@ class MainDrawer extends StatelessWidget {
       DrawerTile(
         title: "Update DB",
         onTap: () async {
-          final BaseInformationDatabase database = BaseInformationDatabase();
-          await database.init();
-          database.updateBaseInformation();
+          database.update();
         },
       ),
       DrawerTile(
         title: "Read DB",
         onTap: () async {
-          final BaseInformationDatabase database = BaseInformationDatabase();
-          await database.init();
-          await database.getBaseInformation(Language.german());
-        },
-      ),
-      DrawerTile(
-        title: "Print DB",
-        onTap: () async {
-          final BaseInformationDatabase database = BaseInformationDatabase();
-          await database.init();
-          await database.printContent();
+          List<PokemonBaseInformation> list =
+              await database.getAll(Language.german());
+          print(list.toString());
         },
       ),
       DrawerTile(
         title: "Search 'Bis'",
         onTap: () async {
-          final BaseInformationDatabase database = BaseInformationDatabase();
-          await database.init();
-          print("Searching for 'Bis'");
-          await database.search("Bis");
+          List<PokemonBaseInformation> list = await database.search("Bis");
+          print(list.toString());
         },
       ),
       DrawerTile(
         title: "Search 'ak'",
         onTap: () async {
-          final BaseInformationDatabase database = BaseInformationDatabase();
-          await database.init();
-          print("Searching for 'ak'");
-          await database.search("ak");
+          List<PokemonBaseInformation> list = await database.search("ak");
+          print(list.toString());
         },
       ),
     ];
